+++ interpolation/interpolationSettings.cpp
@@ -260,32 +260,32 @@ void Crit3DInterpolationSettings::setUseMultipleDetrending(bool newUseMultipleDe
     useMultipleDetrending = newUseMultipleDetrending;
 }
 
-std::vector<double> Crit3DInterpolationSettings::getMultiRegressionSlopes() const
+std::vector<float> Crit3DInterpolationSettings::getMultiRegressionSlopes() const
 {
     return multiRegressionSlopes;
 }
 
-void Crit3DInterpolationSettings::setMultiRegressionSlopes(const std::vector<double> &newMultiRegressionSlopes)
+void Crit3DInterpolationSettings::setMultiRegressionSlopes(const std::vector<float> &newMultiRegressionSlopes)
 {
     multiRegressionSlopes = newMultiRegressionSlopes;
 }
 
-std::vector<double> Crit3DInterpolationSettings::getMultiRegressionAvgs() const
+std::vector<float> Crit3DInterpolationSettings::getMultiRegressionAvgs() const
 {
     return multiRegressionAvgs;
 }
 
-void Crit3DInterpolationSettings::setMultiRegressionAvgs(const std::vector<double> &newMultiRegressionAvgs)
+void Crit3DInterpolationSettings::setMultiRegressionAvgs(const std::vector<float> &newMultiRegressionAvgs)
 {
     multiRegressionAvgs = newMultiRegressionAvgs;
 }
 
-std::vector<double> Crit3DInterpolationSettings::getMultiRegressionStdDevs() const
+std::vector<float> Crit3DInterpolationSettings::getMultiRegressionStdDevs() const
 {
     return multiRegressionStdDevs;
 }
 
-void Crit3DInterpolationSettings::setMultiRegressionStdDevs(const std::vector<double> &newMultiRegressionStdDevs)
+void Crit3DInterpolationSettings::setMultiRegressionStdDevs(const std::vector<float> &newMultiRegressionStdDevs)
 {
     multiRegressionStdDevs = newMultiRegressionStdDevs;
 }
